VENV/Streamlit/bases_streamlit/requirements.txt
matplotlib
numpy
seaborn
panda